fix: completely rework AnythingLLM deployment
The previous incarnation was cribbed from a very dissimilar workshop
environment. To start with, the "primer" was designed to pull images to
nodes to reduce startup times of Workbenches when an attendee started
them. That's not something we're doing here, as any additional
workbenches required will be automated before long as we work out the
details of additional use case examples to scaffold. For this reason,
the workbench image puller was completely removed. This simplified RBAC,
alongside removing unnecessary ClusterRole grants to our user that
applied for different workshop content.

Since we expect to probably deploy more than one AnythingLLM workbench
in the long run, the one deployed here has been renamed to be a bit
more specific to the Aligned Granite model deployed into the cluster.

Additionally, the old Job was unhelpfully named
"create-and-apply-resources" and this non-descriptive name caused some
confusion about what it was doing. Its presence in the "primer"
workspace may have made sense in a multi-user, multi-seat workshop, but
here it just adds additional confusion. The naming and placement of this
Job has been corrected.

Finally, the old content of the Job was a long and difficult to follow
shell script, embedded as a text block, inside the container args for
the PodTemplate in the Job. This made it difficult to follow, impossible
to lint, and generally too complex. This is why it had, for example,
completely unused ServiceAccounts and RoleBindings hanging around in it.
By being a bit more conservative with what we do in the script,
non-declaratively, and then compartmentalizing the script into its own
.sh file, with separate YAML manifest templates that can be linted, we
simplify it greatly.

Future work will likely involve making this Kustomize overlay better
suited for reuse for additional AnythingLLM workbenches, or (more
likely) moving it to a Helm Chart to make it simpler to reuse than
complex sets of patches and multi-layered overlays. The work in
separating the pieces that can remain fully declarative from those that
need to run in a job is portable to both of those approaches, and will
be reusable.